Fix compiler warnings.

+++ src/session/session_api.c
@@ -906,7 +906,7 @@ __session_transaction_sync(WT_SESSION *wt_session, const char *config)
 	 * timeout.
 	 */
 	WT_ERR(__wt_config_gets_def(
-	    session, cfg, "timeout_ms", UINT_MAX, &cval));
+	    session, cfg, "timeout_ms", (int)UINT_MAX, &cval));
 	if ((unsigned int)cval.len != UINT_MAX) {
 		timeout_ms = (uint64_t)cval.val;
 		forever = 0;
